Adds support for Python coveraging.
This extends the "coveraging" option to apply coveraging to any Python
programs executed during the build.  It adds a new option
"python_coverage" to allow configuring the command to be used for
applying coveraging to Python programs, with a default value of
"coverage run" (which is typically appropriate, provided that the
coverage package is installed).  Setting this to the null string
disables Python coveraging.

If "python_coverage" has its default value and parallel builds are
enabled, "--parallel-mode" is automatically appended to the coverage
command.  If a non-default coverage command is used, it's the user's
responsibility to handle parallelism appropriately.

Note that this does not apply coveraging to generated Python code.

This also adds the coverage result files to .gitignore, including the
C coverage files which weren't there previously.

TESTED:
Ran "scons check" as well as sample cases of regress-driver -b and -v,
both with and without coveraging.  Also verified that it works with the
full path to the tool supplied as the python_coverage option.