Improves coverage testing of fake.py.
This adds three "regression" tests which run passthrough.log in all
three transport modes, for more complete coverage in fake.py.  The
target for all three is 'gpsfake-tests'.

Although the original motivation for 'test-nondaemon' was to exclude
slow-running tests, and the gpsfake tests are fairly quick, it would
be misleading to classify them as "nondaemon".  So a new 'test-quick'
target is added which includes these but not the full set of
regression tests.

TESTED:
Ran all relevant test targets, with and without coveraging.