Removes build-time dependency on 'python2'.
Since some systems lack a 'python2' command, the switch to 'python2'
in the shebang lines can cause failures.  This fix gets around it by
explicitly invoking $PYTHON for all Python tools (except valgrind.py,
which already gets the versioned shebang line).  It also exports
$PYTHON to the environment, so that regress-driver can do the same
instead of defaulting to 'python2'.

If scons becomes Python3-compatible before GPSD does, or if GPSD
becomes pickier about Python versions than scons, some additional code
will be needed to find a suitable Python version, but for now just
using the same Python as scons is adequate (and is already what was
being done for some programs, just not all).

This does *not* address the issue of the *installed* Python programs
not working without 'python2'.  That's left for a separate change.

TESTED:
Verified that "scons build-all check" now works without a 'python2' in
the command path, both with and without coveraging.