Fixes gpsprof to work with Python 2.6.
The dict-without-values construct as a set initializer doesn't
work in Python 2.6, though the usual set constructor does.

This is not a recent bug.

TESTED:
Verified that gpsprof now works in Python 2.6, 2.7, 3.3, 3.4, and 3.5,
at least in the TPV mode.  Unable to test the PPS cases since gpsd
doesn't recognize my PPS source (but the syntax is identical).