Fixes client.py bug ausing RTCM3.2 regression-test failures.
When an entire 4K buffer of packet data contains no newlines, it's
simply retained for future use.  But the code was failing to clear the
previous response in that case, causing the caller to see duplicate
data.

Since this bug only mattered with lines longer than 4K bytes, and
since even then it varied as a function of the alignment of lines
vs. 4K boundaries, it was never seen on previous regression tests.

TESTED:
Ran "scons build-all check" with the updated code and the corrected
test data for RTCM3.2.

Signed-off-by: Gary E. Miller <gem@rellim.com>