Fixes fake.py 'packet' import for Python 3.
This changes the import of the 'packet' module to be an explicit
relative import, which is valid in both Python2 and Python 3.  The old
implicit relative import doesn't work in Python 3.

TESTED:
Ran "scons build-all check" (with Python 2.7).  Also verified that the
regression tests and test_maidenhead.py run with Python 2.6, and that
gpsfake -T and test_maidenhead.py run with Python 3 when the
extensions are built for Python 3 (not yet a build option).