Fixes a client.py exception for Python 3.
In Python 3, exception objects are local to the try/except block,
making them inaccessible outside that block.  It was probably actually
an error to set the "message" to be the exception object rather than
its string representation, anyway.

TESTED:
Ran xgps without gpsd running, and verified that the proper error box
is displayed, in both Python 2 and Python 3 (with appropriate builds
of the extensions).