Fixes Gtk warnings related to xgps/xgpsspeed error boxes.
The Gtk MessageDialog function gives a warning if no parent window is
specified.  Since the main window already exists in those cases, it's
simply a matter of passing the extra argument to MessageDialog.

This has the side effect of changing the position of the dialog box,
such that it's centered over the parent window rather than being near
its upper left-hand corner.  In the xgps case, it appears to be
centered over whichever subwindow was most recently updated.

TESTED:
Verified that both the gpsd not running and gpsd stopped running cases
no longer produce warning messages on the terminal.