Changes type checks in misc.py:isotime to use isinstance().
Using isinstance() rather than type() == is preferable for a few
reasons:

1) It's generally considered more pythonic.

2) Referring to types by name is a bit more readable than applying
type() to sample values.

3) Both 'str' and the Python 2 'unicode' type can be checked in a
single call.

4) Since the 'str' type check was the only use of a Unicode literal in
the common code, changing this eliminates Python 3.2 incompatibility.

In Python 2, both 'str' and 'unicode' are derived from the abstract
'basestring' class.  In Python 3, 'unicode' has been folded into
'str', and the latter is itself the base class.  Thus, an import-time
conditional is used to set up the proper STR_CLASS for comparison.

TESTED:
Ran "scons build-all check" (with Python 2.7).  Also verified that the
regression tests and test_maidenhead.py run with Python 2.6, and that
gpsfake -T and test_maidenhead.py run with Python 3.2-3.5 when the
extensions are built for Python 3 (not yet a build option).  Also
verified that xgps and xgpsspeed work with 2.7 and 3.2-3.5 (with
a pending fix to xgps for 3.2).