Removes unnecessary Unicode literal from xgps.
Since the radio-button labels in the add_ui_from_string() text are
just internal labels not visible to the user, there's no reason that
they need to rely on non-USASCII characters.  With that text confined
to 7-bit ASCII, there's no need to use Unicode for the literal, which
in turn eliminates the last incompatibility with Python 3.2.  Note
that this isn't the only case where the internal labels mismatch the
user-visible text (the other ones involve embedded blanks).

The fact that Gtk3 assumes 'ascii' for the encoding in this case may
actually be a bug.

TESTED:
Ran xgps with Python 2.7 and 3.2-3.5 (using appropriately built
extensions).