Fixes gpscat and test_maidenhead.py for Python 3.
This fixes two issues with gpscat in Python 3:

1) The data accumulation buffer needs to be 'bytes' rather than a
'str' in Python 3.  This is easily accomplished by using bytes() as
the initializer/reinitializer.

2) The hexdump routine (which was also iterating in a rather strange
way) needs to obtain int values for the characters of the buffer in a
2/3-independent manner.  This is easily accomplished by converting it
to a bytearray, which works in both Python 2 and Python 3.  This isn't
exactly the same as the Python 3 'bytes' type, since it's
unnecessarily mutable, but it's an easy fix that doesn't require
version conditionals.

Also, test_maidenhead.py works with Python 3 after updating the
libraries, so the shebang line and comments now reflect that.

TESTED:
Ran them under Python 2.6, 2.7, and 3.2-3.5 (with appropriate builds of
the extensions).