Adds "target Python" support to SConstruct.
This adds a target_python option to allow building for a specific
Python version which may differ from the Python running scons.  If
blank (currently the default), it uses the scons Python as before, but
otherwise it's expected to be a command to launch some version of
Python.  In the latter case, all Python-related configuration
parameters are obtained by running the specified Python as an external
program.

The primary effect of this is to allow the C extensions to be built
appropriately for a given Python version, but it also means that all
Python code used for tests is executed with the target Python (which
is necessary in some cases, and desirable for consistency in all
cases).  However, any Python programs executed purely as build helpers
are executed with the scons Python for consistency.

Note that any "pure Python" programs run from the build directory can
be run with any compatible Python version, but programs using the C
extensions will only work with Python versions compatible with the
target_python used when building.  The built C extensions may or may
not use versioned naming, depending on how the target Python specifies
the configuration.

The default value of python_libdir is no longer the libdir for the
scons Python, but instead the default blank value specifies that the
actual value should be obtained from the target Python.  It should
rarely be appropriate to modify this value.

TESTED:
Ran "scons build-all check valgrind-audit" with the default Python
(2.7) and with all six available compatible versions (2.6-2.7 and
3.2-3.5).  Also ran gegps, gpscat, gpsprof, xgps, and xgpsspeed in the
same circumstances.  Except for xgps and xgpsspeed with 2.6, which
doesn't work due to py-gobject3 unavailability for 2.6, everything
worked correctly.