Fixes SConstruct clean/help/minimal bugs from commit 1c26179.
Because the new Python config setup wasn't being run in the cleaning
or helping cases, certain variables that were still being referenced
weren't being defined, causing scons to crash.  There are two
different cases:

1) In the cleaning case, it's actually necessary to run the Python
config code, because some of the filenames to be cleaned may be
affected by the config values.  This reworks things to run that
portion of the config section even in the cleaning case, though with a
slight tweak to make a couple of the messages less verbose.

2) In the helping case, nothing related to building is actually
needed, but providing valid values is easier than fixing all the
places that need them.  To avoid having to run a possible external
target Python, it obtains the values unconditionally from the scons
Python in this case.

Also, the minimal build was broken due to a couple of lines of code
being inappropriately included in the python=no case.  That is now
fixed as well.

TESTED:
Ran "scons build-all check", "scons -c" and "scons -h", with and
without a target_python specified, and with and without minimal=yes.
All cases work, and no "cleaning leaks" were observed.  Also ran the
broken-build checker up through length 1, seeing only the two
long-standing failures, plus two (relatively) new ones that are almost
certainly unrelated to this change.