Fixes integer division for Python 3.
This changes a few integer division cases to use the '//'
floored-division operator, matching the normal Python 2 (C-like) '/'
behavior, for compatibility with the switch to the "mathematical"
divison operator in Python 3 (and in Python 2 with the "future
division" import).  This was more about keeping the behavior the same
than determining whether floored division is actually the proper
choice.

One place where floored division is definitely wanted is in the GPS
week calculations in leapsecond.py, which are now OK for Python 3,
though currently that module is only used by SConstruct, and hence not
with Python 3.

Two other minor fixes:

1) The GPS base date is corrected in the comment in leapsecond.py.

2) The fit_to_grid() function in xgps now consistently returns floats,
rather than returning either ints or floats depending on the line
width.

TESTED:
Ran "scons build-all check" with all 6 supported Python versions.
Also ran xgps and xgpsspeed with all but 2.6.