Fixes leapseconds.py for Python 3.
This fixes two Python 3 issues with leapsecond.py:

1) The urllib module is subdivided in Python 3.  Only the 'request'
submodule is needed here.  Either that or the Python 2 urllib is now
imported as 'urlrequest'.

2) The data returned by urlopen() is 'bytes' in Python 3, requiring
the polystr transformation to be usable as a string.

Also, the (duplicate) isotime() definition here has been updated to
match the version in gps/misc.py.  It is still duplicated to avoid a
dependency on the latter, as is the needed portion of the polystr
code.

TESTED:
Ran "scons build-all check".  Also ran "leapsecond.py -f" with all six
supported Python versions, and verified that the result matches the
existing leapseconds.cache.  Did *not* test the updated retrieve()
function, but it's not used anywhere.