Changes fake.py subclasses to use super() in __init__().
This doesn't really change any behavior, but serves as a better
example for future code.  Note that this can't be done so easily for
gps.gps, due to the mismatched init signatures.

This uses the older super() syntax, which is compatible with both
Python 2 and Python 3.

TESTED:
Ran "scons build-all check" with all six supported Python versions.
Also tested exception objects manually.