Fixes more devtools programs for Python 3.
This fixes the aivdmtable, regressdiff, sizes, and striplog programs
to work with Python 3.  The "striplog" program didn't require any
actual changes, but it still adds the usual compatibility comment and
future import.

In the case of "sizes", it also adjusts the scons cleanup pattern to
include the temporary directory, and adds explicit checking (with an
exception) for build failures.

TESTED:
Ran all four programs with both Python 2 and Python 3.