Adds fallback for target_python handling with older scons.
The CheckProg function is a relatively recent addition to scons, and
is missing from, e.g., scons 2.3.0.  It's worth using it to simplify
the error case, but fallback code is needed to allow building with a
nonempty target_python option and an older scons.  This fix makes it
work correctly in the non-error case, and also avoids crashing the
build in the common error case, albeit without avoiding duplicate
error messages.

TESTED:
Ran "scons -c" with blank, valid, and invalid target_python values
both on OSX with scons 2.4.1 and Linux with scons 2.3.0.