Fixes cycle_analyzer for Python 3 (as well as other bugs).
It appears that this program hadn't been used in a while, since it had
bugs unrelated to Python 3.  After fixing some bugs, it's able to run
successfully on all the test logfiles, except that it chokes on the
binary data in ac12_binary (due to a newline within the binary data),
and on the known malformed sentence in triton400.

In order to allow the utility functions in gps.misc to be imported, it
adds a symlink from devtools/gps to gps.  Due to some weirdness in
Python 3 imports, there didn't seem to be any other way to make this
work with uninstalled libraries, and in any case it allows the local
libraries to take precendence over the system libraries.

TESTED:
Ran with all six supported Python versions on all daemon logfiles,
with only the errors noted above.