Fixes uninstall to remove Python package directory.
Uninstall was removing the contents of the Python package directory,
but not the directory itself.  There's an asymmetry where an
unmentioned directory is created automatically when needed, but not
removed when it's rendered empty.

In Python 2, this is a fairly harmless error, albeit slightly unclean,
since directories lacking an __init__.py file are ignored by the
import logic.  But in Python 3, which allows the init file to be
omitted, leaving the empty directory around causes "import gps" to
apparently succeed and create an empty "namespace" module.  Not only
does this confuse programs attempting to detect import failures, but
it also prevents a future import from working even after adjusting
sys.path, since reimporting an already imported module is a NOP.

The fix is simply to include the directory explicitly in the list for
python_install.  This has no actual effect on the install, but
includes the directory in the uninstall.

TESTED:
Ran install/uninstall and verified that the directory is now removed
by the uninstall.