Fixes gps.gps.gps iterator for Python 3.
Python 3 renames the iterator next() method to __next__().  Without
the new name, attempting to use the object as an iterator doesn't
work.  This didn't surface until trying webgps.py.

For compatibility with both Python 2 and Python 3, both names need to
be provided.  This change renames the existing method to conformto
Python 3 naming, and then adds a wrapper to provide the old name for
Python 2.

TESTED:
Ran webgps.py (with uncommitted fixes) with both Python 2 and Python
3.