Fixes contrib/webgps.py for Python 3.
Since the import fallback hack doesn't work properly in Python 3, this
gets rid of it and adds a symlink to make the 'gps' package directory
directly available here (similar to what was done in devtools).  It
then makes a few minor fixes for Python 3.

The default Python 3 pickle protocol 3 doesn't work in Python 2, but
protocol 2 doesn't work correctly in this case either, so there
doesn't seem to be any way to make the tracks file
backward-compatible.  Thus, it just uses the default protocol and
punts at load time if the load fails.

This change also fixes the reversed skyview, for consistency with
pretty much everything else.

There was one case where it crashed with a "dictionary changed size"
error that appeared to be due to the delete_stale method's deleting a
dictionary entry while scanning it, which is a no no.  This has also
been fixed.

Also adds the result files to .gitignore.

TESTED:
Ran with both Python 2 and Python 3.  This included one run of over
three hours without any more crashes.