Fixes bug in filtering of ID 0 in satellite views.
Because the GSV processing invokes nmeaid_to_prn() before checking for
an apparent ID of zero, it was possible for zero (or blank) IDs for
non-GPS sentences to get offset to nonzero values and hence not
filtered out.  This fix ensures that only nonzero values are adjusted.

Due to this bug, two of the regression tests had incorrect reference
results:

1) The nl2008u receiver includes the new signal ID field, though the
reported value is (incorrectly) zero.  Due to another bug (to be fixed
separately), this is interpreted as an additional ID in each GSV
sentence.  The zero-ID filter excluded this correctly for GPGSV, but
for GLGSV it saw phantom entries for "ID 37".

2) The sl869 receiver reports empty fields for unused satellite slots
in any GSV sentence.  Again, this didn't hurt for GPGSV, and the GLGSV
test data happens to have a multiple of four satellites, but the QZGSV
sentences produced phantom "ID 193" entries.

TESTED:
Ran "scons build-all check", which passed after updating the two
regression cases noted above.