Restore the original output format of ntpshmmon.
Changing the output format without bumping the version number was wrong.

Changing the output format without fixing the documentation and
examples (in INSTALL and elsewhere) was wrong.

Removing the 'sample' keyword at the beginning of the line was wrong.
It broke contrib/ntpshmviz.  More generally, it broke the property
that the format was self-describing and extensible.

Removing the comment leader on the second line was wrong.  That meant
parsers for the format, such as the one in contrib/ntpshmviz, would
try to interpret the header as data and throw errors.