Eliminate the one use of the  portability-challenged alloca()
In general, the use of alloca() is discouraged.  Here, there was
only one use, with a maximum size determinable at compile time,
so the code has been reworked to use an ordinary array (without
relying on the C99 variable-length array feature).  This always
allows for a GPS_PATH_MAX-sized device_name, but the amount of
space needed for that is relatively modest, and the behavior is
more reproducible with a fixed allocation.

TESTED:
Ran "scons build-all check" on OSX.  This code shouldn't be
OS-dependent.

Signed-off-by: Gary E. Miller <gem@rellim.com>