Fixes crash when quitting gpsmon in -a mode.
There were two bugs in the alternate input code used in -a mode:

1) It was using strlen() instead of sizeof() to determine the
available space in the input buffer.  This is uninitialized data,
making the result nondeterministic, though it would return zero
in the case observed.

2) It was failing to handle the NULL return from fgets() correctly,
resulting in a call to do_command() with a garbage argument.  This
caused a segfault.

Also, casting the strlen/sizeof value to int was superfluous.

TESTED:
Typing a "q" in -a mode now outputs a prompt (an incorrect one,
but that's another story) and can then quits without crashing.

Signed-off-by: Gary E. Miller <gem@rellim.com>