Fixes Ill-formed TOFF/PPS packet error in gpsmon.
The code  for reporting the  combined host address and  device address
modifies the  same 'path' field that  it uses to  form the combination
This was resulting in a growing accumulation of device names until the
string became too long.  Although the way this code works could use
some improvement, a cheap fix is to check for the presence of the
device name in the path, and skip it when creating the new one.

TESTED:
On both OSX and Linux, ran "scons build-all check" and tested gpsmon
both with and without the -a option, verifying that it no longer
fails and that the reported "device" is the expected combined form.

Signed-off-by: Gary E. Miller <gem@rellim.com>