Fixes inverted offset signs in gpsmon's PPS lines.
In gpsmon, the PPS offsets reported in the display pane and in the
scrolling packet data has opposite signs.  Since the display version
is consistent with the value reported by the ppswatch utility
(from the pps-tools package), this is assumed to be the correct
version, and the "text" version is corrected to match.

TESTED:
Ran gpsmon on both OSX and Linux, verifying that the two forms
of PPS offset reported by gpsmon and the ppswatch version are all
consistent.

Signed-off-by: Gary E. Miller <gem@rellim.com>