Adds skyview rotation support to xgps and xgpsspeed.
This adds an optional argument to rotate the skyview display, making
it possible to orient it correctly based on the direction one is
actually facing.  The specified heading is positioned at the top.

This edit does not update contrib/webgps.py, which is somewhat
more complicated to fix due to the Javascript involvement.

TESTED:
Ran both programs with and without the -r or --rotate option.

Signed-off-by: Gary E. Miller <gem@rellim.com>