Fixes some build failures of test_gpsmm.
The clock_gettime() function was originally a "second-tier" library
function in Linux, implemented in librt and requiring "-lrt" when
linking.  Later versions moved it to libc, with the "-lrt" being
unnecessary.  The build recipe for test_gpsmm here was failing to
account for the optional need for the "-lrt", causing it to fail
to build on debian Wheezy.

TESTED:
Now passes "scons build-all check" on a Beaglebone Black running
Debian Wheezy, where test_gpsmm formerly failed to build.  Also
retested on OSX 10.9 (where librt doesn't exist) for good measure.