Fixes gratuitous rebuild in non-git directories.
The code that generates the version string for development versions
tries to identify the specific code state in one of two ways.  When
the build is within a git repo, it uses "git describe" output.  When
the build is not within a git repo, it uses the date and time.

Before this change, the applicable date and time was the current date
and time, causing revision.h and everything that depends on it to be
rebuilt every time, even if no sources changed.  This change causes it
to use the latest mtime among the relevant sources instead, thus still
having the desired effect without unnecessary changes in the version
string.

Since git doesn't preserve mtimes, the resulting date/time may reflect
the last checkout of the relevant files rather than the last actual
change, but this is still a vast improvement.  And recent versions of
git avoid clobbering the mtimes of files that are not actually altered
by the checkout.

This change also alters the version format somewhat in this case, both
to incorporate the actual version number, and to eliminate some
truncation which was probably slightly incorrect, anyway.

A sample new format is:

	"3.17~dev-2016-09-24T19:06:46"

TESTED:
Tested both git and non-git cases.  Did not test the non-"dev" case,
but that code path is untouched.