Fixes various flakiness in PPS handling.
The pps_thread_activate() code uses a single structure to pass
information to the monitor thread.  The latter copies the data during
its initialization, so there's no long-term conflict across threads,
but there's no guarantee that the thread runs and gets far enough to
do that before the structure is reused.  Note that this bug is fairly
unlikely to affect a multicore machine unless it's heavily loaded, but
easily causes trouble on a single-core machine.

A variety of strange PPS misbehaviors were observed on a Beaglebone
Black in the presence of this bug, including various nonreproducible
PPS initialization problems and failures, and duplicate PPS reports.

This change uses the pps_thread field as a flag to communicate a "copy
acknowledgment" back to pps_thread_activate(), which waits for it
before proceeding.

TESTED:
Launched several times on a Beaglebone Black with KPPS support, and
no longer observed flaky PPS startup in the logs, as well as no
longer observing duplicate PPS events in gpsmon.  Not tested on
other platforms due to the lack of PPS availability.

Signed-off-by: Gary E. Miller <gem@rellim.com>