Removes two redundant or troublesome compile flags.
The two flags -Wmissing-prototypes and -Wmissing-declarations were
being both conditionally included after a validity test, and then
unconditionally included later.  In cases where they were acceptable,
that resulted in doubly including them, and in cases where they
were unacceptable, that meant breaking the build.

The fix is simply to remove the unconditional adds.  This is a
necessary but not sufficient condition for building with CC=c99
to work.

TESTED:
Ran "scons build-all check" on OSX, and verified that the two
flags were still present in the appropriate commands, but now
only once.  Also verified that building with CC=c99 no longer
fails in a manner related to those flags.