Fixes cleaning bug introduced by LIBPATH change.
Although the LIBPATH override removal wasn't bad per se, it
exacerbated a bug where the standard LIBPATH setup was being bypassed
in the "cleaning or helping" cases, causing many targets to mismatch
for cleanup.  That was missed due to not including a "cleanup test" in
the testing.  The fix is simply to move the LIBPATH setup out of the
conditional area.

TESTED:
Ran scons "build-all check" followed by "scons -c" and then "git clean
-dxf", and verified that now only the usual scons-internal files
remain after cleanup.