Makes MAGIC_HAT kludge optional.
This adds a build option called 'magic_hat', and makes the code for
MAGIC_HAT_GPS and MAGIC_LINK_GPS conditional on it.  It currently
defaults to True on Linux and False otherwise.  It's meaningless on
non-Linux platforms, though at present there are no
conditionally-present options, so this one is also present on all
platforms.

TESTED:
Tested on a Beaglebone Black with suitable symlinks added (and running
a kernel with PPS_CLIENT_KTIMER turned off.  Verified that the kludge
behaves as intended when enabled, and is inoperative when disabled.