Adds Linux "fake PPS" detection and avoidance.
If the Linux kernel is built with CONFIG_PPS_CLIENT_KTIMER=y, then a
synthetic PPS source is added to allow testing without a real PPS
source.  However, this source doesn't even run at exactly 1Hz, so any
attempt to use it for real time synchronization is disastrous.

To make matters worse, this is usually the first PPS driver
configured, causing it to appear as /dev/pps0, which is the implied
PPS device used by the "Magic HAT" kludge.

This change adds detection for the fake source (based on its name),
both to provide a warning if it is configured explicitly, and to skip
over it when applying the "Magic HAT" kludge.

TESTED:
Tested both gpsmon and gpsd on a Beaglebone Black with added symlinks
for testing "Magic HAT".  Did this with kernels with and without the
KTIMER enabled.  Verified that the proper PPS device is selected by
"Magic HAT", and that a warning is generated when expected.