Fixes build problem with magic_hat=yes but no KPPS header.
The MAGIC_HAT code was always pointless without KPPS support, but the
recent change to skip the "fake PPS" device caused it to reference a
function which is not defined without KPPS support.  The fix is to
disable MAGIC_HAT altogether in this case.

Doing this required moving the PPS-related configure checks earlier,
so that the magic_hat option can be turned off before it gets exported
to the config file.

TESTED:
Built (on Ubuntu) both with and without sys/timepps.h, and verified
correct setup in gpsd_config.h, as well as successful builds.  Also
ran "scons check" for good measure (in both cases).