Adds bytes-typed alternative response to Python client.
In Python 2, 'bytes' and 'str' are synonyms, but in Python 3 they are
not.  Since the client may return binary data in some cases, 'bytes'
is a more appropriate data type.  Rather than requiring double
conversions in this case (the socket returns 'bytes'), this adds an
alternative response item which is typed as 'bytes'.

Although the accessor method 'data()' has always existed, it's
shadowed by a name collision in some cases, resulting in the
requirement that the caller access the 'response' item directly (as
noted in the documentation).  Thus, it's not possible to avoid the
conversion to 'str' in the cases where it isn't actually used.

This change doesn't bother to add an accessor method for 'bresponse',
for consistency with the way the API is normally used, though that
should be fixed if this aspect of the API is ever cleaned up.

TESTED:
Ran "scons check" on OSX with all supported versions of Python.