Changes fake.py stream type to 'bytes'.
Given that the data returned bye fake.py may include binary data,
'bytes' is a more appropriate type.  Unlike the client API, where it's
easy to provide both return types, it would be less convenient to do
that here, so 'bytes' is preferred as the one choice (and is
consistent with network sockets).

When sending such data to stdout (or stderr), the straightforward
method is to write to sys.stdXXX.buffer rather than sys.stdXXX.  That
doesn't exist in Python 2, but a helper function is now provided to
get the appropriate 'bytes' stream for stdXXX.

Since no previous *release* of GPSD is compatible with Python 3,
changing this now doesn't break anything based on released code.

TESTED:
Ran "scons check" on OSX with all supported Python versions.