Fixes warning about 64-bit literals on 32-bit platforms.
An unadorned 64-bit literal provokes a width warning on 32-bit
platforms.  Depending on whether the compiler promotes such literals
to a 64-bit type, this case could represent an actual malfunction.
The fix is simply to use the "LL" suffix (actually "ULL" in these
cases).

This also switches to ~0 as the all-ones mask, for better readability.

It also changes a nearby "1LL" to "1ULL" for consistency with its
unsigned context.

TESTED:
Ran "scons build-all check" on four versions of OSX, two versions of
Linux, and FreeBSD, OpenBSD, and NetBSD, with two cases being 32-bit.