Decouples Python selection from SCons Python.
This changes the default value of the target_python option to
'python', i.e. the system default Python.  The former blank default
based the selection on the SCons Python, which is guaranteed not to be
the system default Python if the latter is any Python 3.

TESTED:
Running xgps now works when the default Python is Python 3, without
explicitly specifying target_python.