Fixes Mach port leak in OSX clock_gettime fallback.
The commonly circulated code for nanosecond-resolution time of day in
OSX has a bug.  The port obtained by mach_host_self() is a resource
which needs to be freed after use.  This issue has been observed in
connection with a different Mach port use in:

	 https://codereview.chromium.org/279293002/

It's not known how long one has to run gpsd in order to run afoul of
this bug, or whether it has been observed in practice, but fixing it
proactively is wise.  Note that this code disappears in macOS >=10.12,
once native clock_gettime finally appeared.

Also note that the host port is only needed to *obtain* the clock
port, not to *use* the clock port, so it can be freed prior to
actually reading the clock.

TESTED:
Ran "scons build-all check" on OSX 10.5-10.12.
Ran gpsd for a few hours on OSX 10.9.5 (with one receiver).