Fixes some trouble with gpsmon and no PPS_ENABLE.
Until recently, PPS_ENABLE was always true by default, even on
platforms where it was intended to be forced off.  Fixing this exposed
some issues in gpsmon.  This generally related to confusion between
*displaying* PPS info versus *capturing* it.  Since gpsmon can be run
against a remote target, it should be able to report PPS info even
when running on a host that doesn't support PPS, but the code was
conflating the two issues.

The changes here are:

1) There is now a PPS_DISPLAY_ENABLE, which is always true, and should
be used in place of PPS_ENABLE in code related to displaying PPS info,
to the extent possible.

2) The code related to processing remote TOFF and PPS events now uses
this new conditional.  Aside from restoring the presence of the "PPS
offset" lines in the scrolling data, this also fixes a problem where
the display malfunctioned when this code was absent.  The exact nature
of this dependency hasn't yet been determined, but there's a FIXME to
document the issue.

3) The reporting of PPS offsets in the windowed area still doesn't
work without PPS_ENABLE, due to the entanglement between that code and
the pps_thread code.  There are now a couple of FIXMEs related to that
issue, which is being left for future work.

4) A couple of unnecessary PPS_ENABLE conditionals were eliminated by
making use of the existing conditionally-defined lock/unlock
functions.

TESTED:
Ran gpsmon on OSX 10.9 (where PPS_ENABLE is off) against three
receivers (one local and two remote), with the expected results.  Also
ran it on Ubuntu 14 (where PPS_ENABLE is on) against the same three
receivers (again one local and two remote), again with the expected
results.