Increases MAXCHANNELS in xgps.
The current MAXCHANNELS value in xgps, obtained from gps.py, has been
observed to be inadequate.  The value in gps.py is an outdated value
that was never updated when the C value was changed, but using the new
value of 72 in xgps would result in a huge window.  This change
defines a local MAXCHANNELS in xgps, which does not need to match the
value in gps.py.  It's currently set to 28, based on having observed a
case with 25 visible satellites.  It would be better to make it
dynamic, but that would require reworking some of the Gtk setup code.

TESTED:
Ran against a receiver reporting 25 visible satellites, seeing all
satellites reported and no console error messages.