Initial checkin of new os_compat.c source.
This file combines the code from clock_gettime.c, daemon.c, and strl.c
into a single "OS compatibility" module.  This initial version is just
a verbatim concatenation of those three sources, which is not yet
used.  Editing it for initial use is left to a separate commit for
easier tracking of the edits.

Since there's no "git cat" operation, the history of the code in this
file is not directly available in relation to this file, but can be
found in the histories of its three component sources (even after
their upcoming deletions).

TESTED:
Inapplicable.